requirements.txt
BeautifulSoup==3.2.1
Pillow==2.4.0
PyYAML==3.11
argparse==1.2.1
cssselect==0.9.1
goose-extractor==1.0.18
ipython==2.1.0
jieba==0.32
lxml==3.3.5
ner==0.1
nltk==2.0.4
numpy==1.8.1
pymongo==2.7.1
scikit-learn==0.15.0
scipy==0.14.0
selenium==2.42.1
textblob==0.8.4
wsgiref==0.1.2
